修改权限，主要是domain和资源查询方法

+++ thinkjoy-biz-common/src/main/java/cn/thinkjoy/common/dao/IBaseDAO.java
@@ -295,7 +295,7 @@ public interface IBaseDAO<T extends BaseDomain> {
      */
     public List<Map> queryBySql(@Param("executeSql") String executeSql);
 
-    public int queryBySqlCount(@Param("executeSqlCount") String executeSql);
+    public Long queryBySqlCount(@Param("executeSqlCount") String executeSql);
 
     /**
      * 取得泛型类型
